+++ app/Seggu.Services/DtoMappers/RiskDtoMapper.cs
@@ -10,7 +10,7 @@ namespace Seggu.Services.DtoMappers
         public static RiskCompanyDto GetRiskCompanyDto(Risk r)
         {
             var dto = new RiskCompanyDto();
-            dto.Id = r.Id;
+            dto.Id = (int)r.Id;
             dto.Name = r.Name;
             dto.RiskType = RiskTypeDtoMapper.ToString(r.RiskType);
             dto.CoveragesPacks = r.CoveragesPacks.Select(c => CoveragesPackDtoMapper.GetDto(c)).ToList();
